+++ phenol-core/src/main/java/org/monarchinitiative/phenol/formats/hpo/AssociationType.java
@@ -1,0 +1,5 @@
+package org.monarchinitiative.phenol.formats.hpo;
+
+public enum AssociationType {
+  MENDELIAN, DIGENIC, POLYGENIC
+}