+++ phenol-io/src/main/java/org/monarchinitiative/phenol/io/obo/OboOntologyLoader.java
@@ -14,7 +14,6 @@ import com.google.common.collect.*;
 import org.geneontology.obographs.model.*;
 import org.geneontology.obographs.model.meta.BasicPropertyValue;
 import org.geneontology.obographs.owlapi.FromOwl;
-import org.jgrapht.graph.ClassBasedEdgeFactory;
 import org.jgrapht.graph.DefaultDirectedGraph;
 import org.monarchinitiative.phenol.io.owl.OwlOntologyEntryFactory;
 import org.monarchinitiative.phenol.io.owl.generic.Owl2OboTermFactory;
@@ -208,11 +207,8 @@ public final class OboOntologyLoader {
     } if (rootCandSet.size() > 1 ) {
       TermPrefix prefix;
       Optional<TermId> firstId = rootCandSet.stream().findFirst();
-      if (firstId.isPresent()) {
-        prefix = firstId.get().getPrefix();
-      } else { // this should never happen, but if we cannot find a term, use Owl as the prefix
-        prefix = new TermPrefix("Owl");
-      }
+      // getPrefix should always work actually, but if we cannot find a term for some reason, use Owl as the prefix
+      prefix = firstId.map(TermId::getPrefix).orElseGet(() -> new TermPrefix("Owl"));
       // Assumption: "0000000" is not used for actual terms in any OBO ontology
       rootId = new TermId(prefix,"0000000");
       Term rootTerm = new Term();